refactor: replace validate-peer-dependencies with sermver.peer-range
