+++ hooks/read-package-hook/package.json
@@ -33,7 +33,7 @@
     "@pnpm/parse-overrides": "workspace:*",
     "@pnpm/parse-wanted-dependency": "workspace:*",
     "@pnpm/types": "workspace:*",
-    "@pnpm/validate-peer-dependencies": "workspace:*",
+    "@pnpm/semver.peer-range": "workspace:*",
     "@yarnpkg/extensions": "catalog:",
     "normalize-path": "catalog:",
     "ramda": "catalog:",
