+++ pnpm-workspace.yaml
@@ -32,6 +32,7 @@ packages:
   - releasing/*
   - resolving/*
   - reviewing/*
+  - semver/*
   - store/*
   - text/*
   - workspace/*
