+++ semver/peer-range/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@pnpm/validate-peer-dependencies",
-  "version": "1000.0.0",
+  "name": "@pnpm/semver.peer-range",
+  "version": "1000.0.0-0",
   "description": "Check for invalid peer dependency versions",
   "main": "lib/index.js",
   "types": "lib/index.d.ts",
@@ -12,13 +12,13 @@
     "node": ">=18.12"
   },
   "scripts": {
-    "lint": "eslint \"src/**/*.ts\" \"test/**/*.ts\"",
+    "lint": "eslint \"src/**/*.ts\"",
     "_test": "jest",
-    "test": "pnpm run compile && pnpm run _test",
+    "test": "pnpm run compile",
     "prepublishOnly": "pnpm run compile",
     "compile": "tsc --build && pnpm run lint --fix"
   },
-  "repository": "https://github.com/pnpm/pnpm/blob/main/packages/validate-peer-dependencies",
+  "repository": "https://github.com/pnpm/pnpm/blob/main/semver/peer-range",
   "keywords": [
     "pnpm10",
     "pnpm",
@@ -28,15 +28,13 @@
   "bugs": {
     "url": "https://github.com/pnpm/pnpm/issues"
   },
-  "homepage": "https://github.com/pnpm/pnpm/blob/main/packages/validate-peer-dependencies#readme",
+  "homepage": "https://github.com/pnpm/pnpm/blob/main/semver/peer-range#readme",
   "dependencies": {
-    "@pnpm/error": "workspace:*",
-    "@pnpm/types": "workspace:*",
-    "@types/semver": "catalog:",
     "semver": "catalog:"
   },
   "devDependencies": {
-    "@pnpm/validate-peer-dependencies": "workspace:*"
+    "@pnpm/semver.peer-range": "workspace:*",
+    "@types/semver": "catalog:"
   },
   "funding": "https://opencollective.com/pnpm",
   "exports": {
